refactoring initSections.js
